+++ src/main/java/service/GenreService.java
@@ -1,34 +1,41 @@
 package service;
 
-import aop.Cached;
+import dto.GenreDto;
+import exception.GenreDoesNotExistException;
+import mapper.LibraryMapper;
 import model.Genre;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import repository.GenreDao;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class GenreService {
 
     private final GenreDao genreDao;
+    private final LibraryMapper libraryMapper;
 
-    public GenreService(GenreDao genreDao) {
+    public GenreService(GenreDao genreDao, LibraryMapper libraryMapper) {
         this.genreDao = genreDao;
+        this.libraryMapper = libraryMapper;
     }
 
-    public List<Genre> getAllGenres() {
-        return genreDao.findAll();
+    @Transactional(readOnly = true)
+    public List<GenreDto> getAllGenres() {
+        return libraryMapper.toGenreDtos(genreDao.findAll());
     }
 
-    @Cached
-    public Genre getGenreById(Long id) {
-        try {
-            return genreDao.findById(id);
-        } catch (EmptyResultDataAccessException e){
-            return null;
+    @Transactional(readOnly = true)
+    public Optional<GenreDto> getGenreById(Long id) {
+        Genre genre = genreDao.findById(id);
+
+        if (genre == null) {
+            throw new GenreDoesNotExistException(id);
         }
+
+        return Optional.of(libraryMapper.toGenreDto(genre));
     }
 
     @Transactional
@@ -40,7 +47,11 @@ public class GenreService {
 
     @Transactional
     public void updateGenre(long id, String name) {
-        Genre genre = getGenreById(id);
+        Genre genre = genreDao.findById(id);
+        if (genre == null) {
+            throw new GenreDoesNotExistException(id);
+        }
+
         genre.setName(name);
 
         genreDao.update(genre);
@@ -48,6 +59,9 @@ public class GenreService {
 
     @Transactional
     public void deleteGenre(Long id) {
+        if (genreDao.findById(id) == null) {
+            throw new GenreDoesNotExistException(id);
+        }
         genreDao.delete(id);
     }
 }
